--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/BMpozaFeesManagmentSystemApplication.java
@@ -1,13 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BMpozaFeesManagmentSystemApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BMpozaFeesManagmentSystemApplication.class, args);
-	}
-
-}
