--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/Exceptions/FeeStructureNotFoundException.java
@@ -1,7 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.Exceptions;
-
-public class FeeStructureNotFoundException extends RuntimeException {
-    public FeeStructureNotFoundException(String message) {
-        super(message);
-    }
-}
