--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/Exceptions/StudentNotFoundException.java
@@ -1,7 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.Exceptions;
-
-public class StudentNotFoundException extends RuntimeException{
-    public StudentNotFoundException(String description){
-        super(description);
-    }
-}
