--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/config/SecurityConfig.java
@@ -1,46 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.config;
-
-import com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.service.UserService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@Configuration
-@EnableWebSecurity
-public class SecurityConfig {
-
-    @Autowired
-    private UserService userService;
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        http.
-            csrf(customizer -> customizer.disable()).
-            authorizeHttpRequests(customizer -> customizer.anyRequest().authenticated()).
-            httpBasic(Customizer.withDefaults()).
-            sessionManagement(customizer -> customizer.sessionCreationPolicy(SessionCreationPolicy.STATELESS));
-
-        return http.build();
-
-    }
-
-    @Bean
-    public AuthenticationProvider provider(){
-        DaoAuthenticationProvider provider = new DaoAuthenticationProvider();
-
-        provider.setUserDetailsService(userService);
-        provider.setPasswordEncoder(new BCryptPasswordEncoder(12));
-
-        return provider;
-    }
-
-
-}
