--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/model/FeePayment.java
@@ -1,91 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.model;
-
-import jakarta.persistence.*;
-
-import java.math.BigDecimal;
-import java.time.LocalDateTime;
-
-@Entity
-public class FeePayment {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "invoice_id", nullable = false)
-    private Invoice invoice;
-
-    @ManyToOne
-    @JoinColumn(name = "student_id", nullable = false)
-    private Student student;
-
-    private BigDecimal amountPaid;
-
-    private LocalDateTime paymentDate = LocalDateTime.now();
-
-    @Enumerated(EnumType.STRING)
-    @Column(length = 50)
-    private PaymentMethod paymentMethod;
-
-    // No-args constructor
-    public FeePayment() {}
-
-    // All-args constructor
-    public FeePayment(Invoice invoice, Student student, BigDecimal amountPaid, LocalDateTime paymentDate, PaymentMethod paymentMethod) {
-        this.invoice = invoice;
-        this.student = student;
-        this.amountPaid = amountPaid;
-        this.paymentDate = paymentDate;
-        this.paymentMethod = paymentMethod;
-    }
-
-    // Getters and Setters
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Invoice getInvoice() {
-        return invoice;
-    }
-
-    public void setInvoice(Invoice invoice) {
-        this.invoice = invoice;
-    }
-
-    public Student getStudent() {
-        return student;
-    }
-
-    public void setStudent(Student student) {
-        this.student = student;
-    }
-
-    public BigDecimal getAmountPaid() {
-        return amountPaid;
-    }
-
-    public void setAmountPaid(BigDecimal amountPaid) {
-        this.amountPaid = amountPaid;
-    }
-
-    public LocalDateTime getPaymentDate() {
-        return paymentDate;
-    }
-
-    public void setPaymentDate(LocalDateTime paymentDate) {
-        this.paymentDate = paymentDate;
-    }
-
-    public PaymentMethod getPaymentMethod() {
-        return paymentMethod;
-    }
-
-    public void setPaymentMethod(PaymentMethod paymentMethod) {
-        this.paymentMethod = paymentMethod;
-    }
-}
