--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/model/InvoiceStatus.java
@@ -1,7 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.model;
-
-public enum InvoiceStatus {
-    Pending,
-    Paid,
-    Overdue
-}
