--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/repository/FeePaymentRepository.java
@@ -1,7 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.repository;
-
-import com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.model.FeePayment;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface FeePaymentRepository extends JpaRepository<FeePayment,Long> {
-}
