--- src/main/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/repository/StudentRepository.java
@@ -1,7 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.repository;
-
-import com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System.model.Student;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface StudentRepository extends JpaRepository<Student,Long> {
-}
