+++ src/main/java/com/onyx/Exceptions/FeeStructureNotFoundException.java
@@ -1,0 +1,7 @@
+package com.onyx.Exceptions;
+
+public class FeeStructureNotFoundException extends RuntimeException {
+    public FeeStructureNotFoundException(String message) {
+        super(message);
+    }
+}
