+++ src/main/java/com/onyx/Exceptions/InvoiceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.onyx.Exceptions;
+
+public class InvoiceNotFoundException extends RuntimeException{
+    public InvoiceNotFoundException(String description){
+        super(description);
+    }
+}
