+++ src/main/java/com/onyx/Exceptions/StudentNotFoundException.java
@@ -1,0 +1,7 @@
+package com.onyx.Exceptions;
+
+public class StudentNotFoundException extends RuntimeException{
+    public StudentNotFoundException(String description){
+        super(description);
+    }
+}
