+++ src/main/java/com/onyx/OnyxApplication.java
@@ -1,0 +1,13 @@
+package com.onyx;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class OnyxApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(OnyxApplication.class, args);
+	}
+
+}
