+++ src/main/java/com/onyx/model/FeeStructure.java
@@ -1,0 +1,55 @@
+package com.onyx.model;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+
+import java.math.BigDecimal;
+
+@Entity
+public class FeeStructure {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private int grade;
+
+    private BigDecimal totalFee;
+
+    // No-args constructor
+    public FeeStructure() {}
+
+    // All-args constructor
+    public FeeStructure(int grade, BigDecimal totalFee) {
+        this.grade = grade;
+        this.totalFee = totalFee;
+    }
+
+    // Getters and Setters
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public int getGrade() {
+        return grade;
+    }
+
+    public void setGrade(int grade) {
+        this.grade = grade;
+    }
+
+    public BigDecimal getTotalFee() {
+        return totalFee;
+    }
+
+    public void setTotalFee(BigDecimal totalFee) {
+        this.totalFee = totalFee;
+    }
+}
