+++ src/main/java/com/onyx/model/InvoiceStatus.java
@@ -1,0 +1,7 @@
+package com.onyx.model;
+
+public enum InvoiceStatus {
+    Pending,
+    Paid,
+    Overdue
+}
