+++ src/main/java/com/onyx/repository/FeePaymentRepository.java
@@ -1,0 +1,7 @@
+package com.onyx.repository;
+
+import com.onyx.model.FeePayment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface FeePaymentRepository extends JpaRepository<FeePayment,Long> {
+}
