+++ src/main/java/com/onyx/repository/StudentRepository.java
@@ -1,0 +1,7 @@
+package com.onyx.repository;
+
+import com.onyx.model.Student;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface StudentRepository extends JpaRepository<Student,Long> {
+}
