+++ src/main/java/com/onyx/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.onyx.repository;
+
+import com.onyx.model.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UserRepository extends JpaRepository<User,Long> {
+    Optional<User> findByUsername(String username);
+}
