+++ src/main/java/com/onyx/service/UserService.java
@@ -1,0 +1,25 @@
+package com.onyx.service;
+
+import com.onyx.model.User;
+import com.onyx.model.UserPrincipal;
+import com.onyx.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class UserService implements UserDetailsService {
+
+    @Autowired
+    private UserRepository userRepo;
+
+    @Override
+    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+
+        User user = userRepo.findByUsername(username).orElseThrow(()-> new UsernameNotFoundException("Username Not Found"));
+        return new UserPrincipal(user);
+    }
+}
