--- src/test/java/com/kingjnr/feesManagementSystem/B_Mpoza_Fees_Managment_System/BMpozaFeesManagmentSystemApplicationTests.java
@@ -1,13 +1,0 @@
-package com.kingjnr.feesManagementSystem.B_Mpoza_Fees_Managment_System;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class BMpozaFeesManagmentSystemApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
