+++ src/test/java/com/onyx/BMpozaFeesManagmentSystemApplicationTests.java
@@ -1,0 +1,13 @@
+package com.onyx;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BMpozaFeesManagmentSystemApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
